test-pack/requirements.txt
pytest==7.4.0
pytest-bdd==6.1.1
pytest-asyncio==0.21.1
paho-mqtt==1.6.1
requests==2.31.0
pytest-mock==3.11.1
